Language and theme

+++ src/Main/Main.Domain/entities/item/ItemFieldValue.cs
@@ -13,7 +13,7 @@ namespace Main.Domain.entities.item
 
         public string? TextValue { get; set; }
         public string? MultilineTextValue { get; set; }
-        public decimal? NumberValue { get; set; }
+        public double? NumberValue { get; set; }
         public string? FileUrl { get; set; }
         public bool? BooleanValue { get; set; }
